requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
openai==1.44.0
notion-client==2.2.1
google-cloud-storage==2.18.2

